requirements.txt
# Django Framework
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0

# Azure Services
azure-ai-documentintelligence==1.0.0b1
azure-core==1.29.5
azure-storage-blob==12.19.0
openai==1.6.1  # For Azure OpenAI

# Async Task Processing
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Development
ipython==8.18.1
